requirements.txt
cython
django>=1.7,<1.8
dj-database-url
dj-static>=0.0.6
django-absolute
django-appconf
django-ghostdown
django-grappelli
mistune
psycopg2
pystache
PyYAML
six
waitress
